abstracted cleanStyleAttributes and getOrganizedAttributes functions
