+++ src/extensions/copy-paste/index.js
@@ -1,0 +1,2 @@
+export { default as getOrganizedAttributes } from './getOrganizedAttributes';
+export { default as cleanStyleAttributes } from './cleanStyleAttributes';
